app/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
python-dotenv
faiss-cpu
numpy
